pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "taskhub"
version = "0.1.0"
description = "Task management tool, supporting import/export from/to different services, with multiple interfaces."
license = "ISC"
authors = ["Timothée Mazzucotelli <pawamoy@pm.me>"]
readme = 'README.md'
repository = "https://github.com/pawamoy/taskhub"
homepage = "https://github.com/pawamoy/taskhub"
keywords = ['tasks', 'tasks-management', 'bugs']
packages = [ { include = "taskhub", from = "src" } ]
include = [
    "README.md",
    "pyproject.toml"
]

[tool.poetry.dependencies]
python = "~3.6"
django = "^2.1"
toml = {version = "^0.10.0",optional = true}
pygithub = {version = "^1.43",optional = true}
pyyaml = {version = "^3.13",optional = true}
xmltodict = {version = "^0.11.0",optional = true}
taskw = {version = "^1.2.0",optional = true}

[tool.poetry.extras]
github = ["pygithub"]
yaml = ["pyyaml"]
toml = ["toml"]
taskwarrior = ["taskw"]
all = ["pygithub", "pyyaml", "toml", "taskw"]

[tool.poetry.dev-dependencies]
ipython = "^7.2"
pytest = "^4.0"
pytest-cov = "^2.6"
pytest-sugar = "^0.9.2"

[tool.poetry.scripts]
taskhub = 'taskhub.cli:main'

[tool.black]
line-length = 120
